+++ tests/testthat/test-functional_testing.R
@@ -54,7 +54,7 @@ local({
       )
     )),
     filter_key = "car",
-      use_dataset_filter = FALSE
+      enable_dataset_filter = FALSE
   )
 
 
@@ -180,7 +180,7 @@ local({
           )
         ),
         filter_key = "car",
-      use_dataset_filter = FALSE
+      enable_dataset_filter = FALSE
       )